Move HTTP request behavior to its own file
