+++ spec/support/http/shared_examples.rb
